everything but constants
