+++ ui/src/components/flows/tasks/TaskObject.vue
@@ -1,107 +1,27 @@
 <template>
     <el-form label-position="top" class="w-100">
         <template v-if="sortedProperties">
-            <template v-for="[key, schema] in requiredProperties" :key="key">
+            <template v-for="[key, fieldSchema] in requiredProperties" :key="key">
                 <template v-if="key === 'id' || isNestedProperty(key)">
-                    <el-form-item :required="isRequired(key)">
-                        <template #label>
-                            <div class="inline-wrapper">
-                                <div class="inline-start">
-                                    <TaskLabelWithBoolean
-                                        :type="getType(schema)"
-                                        :is-boolean="isBoolean(schema)"
-                                        :component-props="componentProps(key, schema)"
-                                    />
-                                    <span v-if="getKey(key)" class="label">
-                                        {{ getKey(key) }}
-                                    </span>
-                                </div>
-                                <el-tag
-                                    v-if="!isAnyOf(schema)"
-                                    disable-transitions
-                                    size="small"
-                                    class="type-tag"
-                                >
-                                    {{ getType(schema) }}
-                                </el-tag>
-                                <el-tooltip
-                                    v-if="!isAnyOf(schema) && hasTooltip(schema)"
-                                    :persistent="false"
-                                    :hide-after="0"
-                                    effect="light"
-                                    placement="left-start"
-                                    :show-arrow="false"
-                                    popper-class="singleton-tooltip"
-                                >
-                                    <template #content>
-                                        <markdown
-                                            class="markdown-tooltip"
-                                            :source="helpText(schema)"
-                                        />
-                                    </template>
-                                    <help />
-                                </el-tooltip>
-                            </div>
-                        </template>
-                        <component
-                            v-if="!isBoolean(schema)"
-                            :is="`task-${getType(schema, key)}`"
-                            v-bind="{...componentProps(key, schema)}"
-                            class="mt-1 mb-2 wrapper"
-                        />
-                    </el-form-item>
+                    <TaskObjectField
+                        v-model="modelValue[key]"
+                        :field-key="key"
+                        :schema="fieldSchema"
+                        :definitions
+                        :task="modelValue"
+                    />
                 </template>
 
                 <template v-else>
                     <TaskWrapper :merge>
                         <template #tasks>
-                            <el-form-item :required="isRequired(key)">
-                                <template #label>
-                                    <div class="inline-wrapper">
-                                        <div class="inline-start">
-                                            <TaskLabelWithBoolean
-                                                :type="getType(schema)"
-                                                :is-boolean="isBoolean(schema)"
-                                                :component-props="componentProps(key, schema)"
-                                            />
-                                            <span v-if="getKey(key)" class="label">
-                                                {{ getKey(key) }}
-                                            </span>
-                                        </div>
-                                        <el-tag
-                                            v-if="!isAnyOf(schema)"
-                                            disable-transitions
-                                            size="small"
-                                            class="type-tag"
-                                        >
-                                            {{ getType(schema) }}
-                                        </el-tag>
-                                        <el-tooltip
-                                            v-if="!isAnyOf(schema) && hasTooltip(schema)"
-                                            :persistent="false"
-                                            :hide-after="0"
-                                            effect="light"
-                                            placement="left-start"
-                                            :show-arrow="false"
-                                            popper-class="singleton-tooltip"
-                                        >
-                                            <template #content>
-                                                <markdown
-                                                    class="markdown-tooltip"
-                                                    :source="helpText(schema)"
-                                                />
-                                            </template>
-                                            <help />
-                                        </el-tooltip>
-                                    </div>
-                                </template>
-                                <component
-                                    v-if="!isBoolean(schema)"
-                                    :is="`task-${getType(schema, key)}`"
-                                    v-bind="{...componentProps(key, schema)}"
-                                    class="mt-1 mb-2 wrapper"
-                                />
-                            </el-form-item>
+                            <TaskObjectField
+                                v-model="modelValue[key]"
+                                :field-key="key"
+                                :schema="fieldSchema"
+                                :definitions
+                                :task="modelValue"
+                            />
                         </template>
                     </TaskWrapper>
                 </template>
@@ -109,56 +29,16 @@
 
             <el-collapse v-model="activeNames" v-if="optionalProperties?.length" class="collapse">
                 <el-collapse-item name="optional" :title="$t('no_code.sections.optional')">
-                    <template v-for="[key, schema] in optionalProperties" :key="key">
+                    <template v-for="[key, fieldSchema] in optionalProperties" :key="key">
                         <TaskWrapper>
                             <template #tasks>
-                                <el-form-item :required="isRequired(key)">
-                                    <template #label>
-                                        <div class="inline-wrapper">
-                                            <div class="inline-start">
-                                                <TaskLabelWithBoolean
-                                                    :type="getType(schema)"
-                                                    :is-boolean="isBoolean(schema)"
-                                                    :component-props="componentProps(key, schema)"
-                                                />
-                                                <span v-if="getKey(key)" class="label">
-                                                    {{ getKey(key) }}
-                                                </span>
-                                            </div>
-                                            <el-tag
-                                                v-if="!isAnyOf(schema)"
-                                                disable-transitions
-                                                size="small"
-                                                class="type-tag"
-                                            >
-                                                {{ getType(schema) }}
-                                            </el-tag>
-                                            <el-tooltip
-                                                v-if="!isAnyOf(schema) && hasTooltip(schema)"
-                                                :persistent="false"
-                                                :hide-after="0"
-                                                effect="light"
-                                                placement="left-start"
-                                                :show-arrow="false"
-                                                popper-class="singleton-tooltip"
-                                            >
-                                                <template #content>
-                                                    <markdown
-                                                        class="markdown-tooltip"
-                                                        :source="helpText(schema)"
-                                                    />
-                                                </template>
-                                                <help />
-                                            </el-tooltip>
-                                        </div>
-                                    </template>
-                                    <component
-                                        :is="`task-${getType(schema, key)}`"
-                                        v-bind="{...componentProps(key, schema)}"
-                                        v-if="!isBoolean(schema)"
-                                        class="mt-1 mb-2 wrapper"
-                                    />
-                                </el-form-item>
+                                <TaskObjectField
+                                    v-model="modelValue[key]"
+                                    :field-key="key"
+                                    :schema="fieldSchema"
+                                    :definitions
+                                    :task="modelValue"
+                                />
                             </template>
                         </TaskWrapper>
                     </template>
@@ -167,56 +47,16 @@
 
             <el-collapse v-model="activeNames" v-if="deprecatedProperties?.length" class="collapse">
                 <el-collapse-item name="deprecated" :title="$t('no_code.sections.deprecated')">
-                    <template v-for="[key, schema] in deprecatedProperties" :key="key">
+                    <template v-for="[key, fieldSchema] in deprecatedProperties" :key="key">
                         <TaskWrapper>
                             <template #tasks>
-                                <el-form-item :required="isRequired(key)">
-                                    <template #label>
-                                        <div class="inline-wrapper">
-                                            <div class="inline-start">
-                                                <TaskLabelWithBoolean
-                                                    :type="getType(schema)"
-                                                    :is-boolean="isBoolean(schema)"
-                                                    :component-props="componentProps(key, schema)"
-                                                />
-                                                <span v-if="getKey(key)" class="label">
-                                                    {{ getKey(key) }}
-                                                </span>
-                                            </div>
-                                            <el-tag
-                                                v-if="!isAnyOf(schema)"
-                                                disable-transitions
-                                                size="small"
-                                                class="type-tag"
-                                            >
-                                                {{ getType(schema) }}
-                                            </el-tag>
-                                            <el-tooltip
-                                                v-if="!isAnyOf(schema) && hasTooltip(schema)"
-                                                :persistent="false"
-                                                :hide-after="0"
-                                                effect="light"
-                                                placement="left-start"
-                                                :show-arrow="false"
-                                                popper-class="singleton-tooltip"
-                                            >
-                                                <template #content>
-                                                    <markdown
-                                                        class="markdown-tooltip"
-                                                        :source="helpText(schema)"
-                                                    />
-                                                </template>
-                                                <help />
-                                            </el-tooltip>
-                                        </div>
-                                    </template>
-                                    <component
-                                        :is="`task-${getType(schema, key)}`"
-                                        v-bind="{...componentProps(key, schema)}"
-                                        v-if="!isBoolean(schema)"
-                                        class="mt-1 mb-2 wrapper"
-                                    />
-                                </el-form-item>
+                                <TaskObjectField
+                                    v-model="modelValue[key]"
+                                    :field-key="key"
+                                    :schema="fieldSchema"
+                                    :definitions
+                                    :task="modelValue"
+                                />
                             </template>
                         </TaskWrapper>
                     </template>
@@ -242,14 +82,9 @@
 
 <script>
     import Task from "./Task";
-    import Information from "vue-material-design-icons/InformationOutline.vue";
-    import Help from "vue-material-design-icons/Information.vue";
-    import Kicon from "../../Kicon.vue";
-    import Editor from "../../inputs/Editor.vue";
-    import Markdown from "../../layout/Markdown.vue";
     import TaskDict from "./TaskDict.vue";
     import TaskWrapper from "./TaskWrapper.vue";
-    import TaskLabelWithBoolean from "./TaskLabelWithBoolean.vue";
+    import TaskObjectField from "./TaskObjectField.vue";
 
     function sortProperties(properties, required) {
         if (!properties) {
@@ -296,13 +131,8 @@
         mixins: [Task],
         components: {
             TaskDict,
-            Information,
-            Help,
-            Kicon,
-            Editor,
-            Markdown,
             TaskWrapper,
-            TaskLabelWithBoolean,
+            TaskObjectField,
         },
         props: {
             properties: {
@@ -333,54 +163,9 @@
             },
         },
         methods: {
-            componentProps(key, schema){
-                return {
-                    modelValue: this.modelValue?.[key],
-                    task: this.modelValue,
-                    "onUpdate:modelValue": (event) => this.onObjectInput(key, event),
-                    root: this.getKey(key),
-                    schema: schema,
-                    required: this.isRequired(key),
-                    definitions: this.definitions
-                }
-            },
-            onObjectInput(propertyName, value) {
-                const currentValue = this.modelValue || {};
-                currentValue[propertyName] = value;
-                this.onInput(currentValue);
-            },
-            isValidated(key) {
-                return (
-                    this.isRequired(key) &&
-                    !this.modelValue?.[key] &&
-                    this.schema.properties[key].default === undefined
-                );
-            },
-            hasTooltip(schema) {
-                return schema.title || schema.description;
-            },
-            helpText(schema) {
-                return (
-                    (schema.title ? "**" + schema.title + "**" : "") +
-                    (schema.title && schema.description ? "\n" : "") +
-                    (schema.description ? schema.description : "")
-                );
-            },
-            isAnyOf(schema) {
-                return !!schema?.anyOf;
-            },
-            isBoolean(schema) {
-                return this.getType(schema) === "boolean";
-            },
             isNestedProperty(key) {
                 return key.includes(".") ||
                     ["interval", "maxInterval", "minInterval", "type"].includes(key);
-            },
-            getKey(key) {
-                if (this.isNestedProperty(key) || key === "id") {
-                    return key;
-                }
-                return key.charAt(0).toUpperCase() + key.slice(1);
             },
         },
     };
@@ -401,60 +186,4 @@
         max-width: 300px !important;
         background: var(--ks-tooltip-background);
     }
-</style>
-
-<style lang="scss" scoped>
-@import "../../code/styles/code.scss";
-
-.el-form-item {
-    width: 100%;
-
-    > :deep(.el-form-item__label) {
-        width: 100%;
-        display: flex;
-        align-items: center;
-        padding: 0;
-    }
-}
-
-.inline-wrapper {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    min-width: 0;
-
-    .inline-start {
-        display: flex;
-        align-items: center;
-        gap: 0.5rem;
-        min-width: 0;
-        flex: 1 1 auto;
-    }
-
-    .label {
-        color: var(--ks-content-primary);
-        min-width: 0;
-        flex: 1;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        font-weight: 600;
-    }
-
-    .type-tag {
-        background-color: var(--ks-tag-background-active);
-        color: var(--ks-tag-content);
-        font-size: 12px;
-        line-height: 20px;
-        padding: 0 8px;
-        padding-bottom: 2px;
-        border-radius: 8px;
-        text-transform: capitalize;
-    }
-
-    .information-icon {
-        color: var(--ks-content-secondary);
-        cursor: pointer;
-    }
-}
 </style>
